Fill in the mutations
The actual mechanics of the game is moving cards between the deck, the
discard pile, players' hands, and players' board. This is captured in
the CardMovement service object.

Each GraphQL type class has class methods for converting DB model
objects into GraphQL representations.

There are still many TODOs unfinished in here, including the entire
backing database for the Player model.

The cards are hardcoded into the Card class. A card ID is the card name
and an autoincrementing number. This needs to be unique so we can move
specific card objects around. A game consists of moving these 52 objects
between all the places it can be.

Since a Game can exist without any players -- it is created, and then
you connect to it -- we need an additional state to represent the
differences between created (zero players), waiting (one player), and
ongoing (two players).