Replace GraphQL with HTML
The vision I had early on was a GraphQL-backed API serving a React app
written in TypeScript. But as I went down that path I realized just how
overkill that is.

A turn consists of taking a card from one pile and moving it into a
another. This can most easily be represented as a form with two radio
sets: one for selecting among the visible source cards, and another for
selecting a destination pile. The hard work is to then manage the shared
backend state of tracking the card into the right pile.

We can represent that in a HTML form most easily, so we do that. I now
have a playable game.

From here we can enhance it with JavaScript, in two directions:

- Use ActionCable to auto-publish backend changes to other players,
using JS to automatically update their DOM.
- Add JS to  make it so dragging a card onto a pile selects the
appropriate radio buttons and submits the movement.

These enhancements will make the game fun to play, but the game is
currently playable with zero JavaScript.