Live updates of the game
This uses ActionCable to maintain a Websocket connection between the
client and server so we can stream updates to the game.

Pass the `connection_secret` as a Websocket param so we can intercept it
during connection and use it to look up the user to be `identified_by`.

The payload is the source pile, destination pile, rich details on the
card itself (if visible to the user in the destination), and any extras.
Extras can include the card ID (useful for when the card details are not
visible, such as the deck or opponent's hand) or the card ID for the top
of the deck.

Additionally in the payload we include the source and destination pile
types -- hand, board, deck, discard. This makes the JavaScript easier to
work with since it never cares about the details of splitting the player
details out.

All the card IDs are exposed, including masked cards and the top of the
deck. This is ostenibly fine except the card ID currently contains the
card name. In the future we'll need to switch to a more obscure card ID.

The JavaScript works by re-arranging the DOM and building new "card"
elements on demand. This means there is duplication between the
`buildCardElement` function and `app/views/cards/_card.html.erb`. Also a
thing to solve in the future.