Encrypt the card ID
The card ID had been in the form of "#{idx}_#{card_name}", with idx
ranging from 0 to 51. This allowed each card to have a unique ID per
game, and also allows us to identify a card given its ID with a simple
string split on `_`.

However, this exposed the card name in plaintext, which meant it was
useless for masked cards such as the top of the deck or the other
players' hands.

Instead, use the `ActiveSupport::MessageEncryptor` class to encrypt and
decrypt using AES. The plaintext card ID can remain the same, but now
it's encrypted.

We need the salt, so pass that as part of the card ID.

The performance is brutally bad when working with 52 cards. A lot of
this is in setting up and tearing down all the objects required. For
example, `Game#set_up_game`, which generates 52 encrypted IDs, took over
three seconds. Similarly, `Game#deck` took just as long to decrypt all
52 card IDs.

Switching to memoization makes it near instantaneous. This means we
re-use the salt for every card in the game, which possibly means someone
could break it if they tried. Hopefully a game will end before
someone cracks it.