Limit the drop targets
This is a complete re-write of the drag-and-drop functionality. Instead
of connecting every droppable and using a 500 to indicate that it wasn't
right, we can instead designate specific piles as droppable and clear
this after each drag end.

Add some CSS to make it more clear. On drag start we highlight the valid
piles in green. On drag over give it a grey background just so you can
see that you've done something. The whole while, make the original card
half transparent.