Submit the form in the background
Use rails-ujs to fire a `submit` action for the form, and add
`data-remote="true"` to the form. Combined, this will submit the form
using AJAX.

Now we need to show the latest. Publish the websocket action updates to
every player. Mask out the cards for all but our player when the
destination pile is a hand. This masking is the bulk of changes in here.