Log errors from CardMovementsController
Now that we are using AJAX for CardMovements form, we are no longer
checking the return value, and therefore no longer showing error
messages. As a quick hack, log the error to the Rails log. Long term we
should handle the AJAX response and also potentially render a different
value based on the response type.