Pilfer through the discard
Expose the ability for a user to move an arbitrary card from the discard
pile into their hand. This is needed for the "Wormhole" mechanic.

We already had this controller in place but it was broken. Fix it up and
link to it.

Add some JavaScript to hide the controls and make it such that clicking
on a card moves it (select the radio and submit the form). The non-JS
fallback will continue to work.