Remove the float CSS
The goal is to show some card listings in a horizontal spread. Use
`display: inline-block` for this instead of `float: left` so we can
avoid going down a floathole.

This is an attempt to fix a CSS bug that I saw in a game with Rebecca
that I cannot reproduce now, where the "Your board" section covers the
cards in your hand.