Show a masked card for the deck
Change the way we render both the deck and the discard pile to be a
"card count": show a pseudo card with a size, then offset a card atop
it. This has a nice effect where you can see the top card (masked, for
the deck) and then below that is a second card-like thing peaking out
with the number of cards in the pile.

We want to handle the case where the discard or deck are empty. We do
this by grabbing any existing card, removing it if it exists, then
adding our new card.

Along the way I learned some DOM:

- [`ChildNode.remove()`] is the new `c.parentElement.removeChild(c)`.
- [`HTMLCollection`] can index by element `name` in addition to treating
it like an array. Turns out this is how
`document.forms["card_movement"]` works.

[`ChildNode.remove()`]: https://developer.mozilla.org/en-US/docs/Web/API/ChildNode/remove
[`HTMLCollection`]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLCollection