Clone the Card when finding it
When we find a Card by ID, we pull it out of a Hash that is computed
statically at file load. We then set the ID. This caused _all_
occurances of that Card -- across all games, and within the same game --
to have the ID that was last set.

Fix that by cloning the Card.